apps/agents/requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-dotenv==1.0.1

# Google Cloud & Vertex AI
google-cloud-aiplatform==1.68.0
google-generativeai==0.8.1
vertexai==1.68.0

# HTTP requests
requests==2.32.3
httpx==0.27.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0

# Utilities
tenacity==9.0.0
pydantic-settings==2.5.0
